# Optimize generated ToString compilation performance