By not calling std::string constructor on every struct field ever.

generated_vulkan_struct_to_string: 79s -> 65s
generated_vulkan_struct_decoders_to_string.cpp: 73s -> 57s