@ishitatsuyuki 2022-11-14
@andrew-lunarg Mind taking a look? This should be fairly trivial to review.